Create gap_analysis.r
It computes the number of taxa represented in a reserve network for a given conservation target and assesses whether this level of representation is significantly lower or greater than expected by chance. An analysis for multiple thresholds to consider a grid cell as protected is undertaken. The inputs are: (1) matrix of presence/absence data by species and area; (2) a table with area codes and the percentage of overlap with protected areas for each area.